requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
boto3==1.4.5
botocore==1.5.92
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.32.0
django-environ==0.4.4
django-extensions==1.8.1
django-s3direct==1.0.3
django-storages==1.6.5
django-webpack-loader==0.5.0
docutils==0.13.1
enum34==1.1.6
futures==3.1.1
google-api-python-client==1.6.5
httplib2==0.10.3
idna==2.5
isort==4.2.15
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==4.1.2
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDrive==1.3.1
pylint==1.7.2
python-dateutil==2.6.1
python-openid==2.2.5
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.18.2
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.10
singledispatch==3.4.0.3
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.10
